examples/mqtt-example/requirements.txt
paho-mqtt
python-qpid-proton